// angular/src/app/share/model/conta.ts
export interface Conta {

  email: string | null;
  password: string | null;
  nome: string | null;
  dataNascimento: Date | null;
  cpf: string | null;
  celular: string | null;
  sexo: string | null;
  cep: string | null;
  pais: string | null;
  estado: string | null;
  cidade: string | null;
  bairro: string | null;
  endereco: string | null;
  complemento: string | null;
}